requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.10
PyJWT==2.10.0
python-decouple==3.8
pytz==2024.2
requests==2.32.3
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
